feat: implement outcome handler and configuration for Rock Paper Scissors game
