+++ exercise/ts/day18/src/settings/conf.ts
@@ -1,0 +1,22 @@
+export const conf = {
+  Rock: {
+    Scissors: "crushes",
+    Lizard: "crushes",
+  },
+  Paper: {
+    Rock: "covers",
+    Spock: "disproves",
+  },
+  Scissors: {
+    Paper: "cuts",
+    Lizard: "decapitates",
+  },
+  Spock: {
+    Rock: "vaporizes",
+    Scissors: "smashes",
+  },
+  Lizard: {
+    Paper: "eats",
+    Spock: "poisons",
+  },
+} as const;
